--- PacMan/main.py
@@ -1,10 +1,0 @@
-from PacMan.game import pacman
-
-def play_game():
-    """
-    
-    """
-    pacman.start_game()
-
-if __name__ == "__main__":
-    play_game()
