--- PacMan/src/rl/__init__.py
@@ -1,1 +1,0 @@
-__all__ = ["Environment", "Graph", "Node", "States", "State"]
